[UPDATE] Proteccion de rutas

+++ Frontend/src/app/shared/toolbar/toolbar.component.ts
@@ -1,4 +1,11 @@
-import { Component, OnInit, AfterViewInit, Input, Output, EventEmitter } from '@angular/core';
+import {
+  Component,
+  OnInit,
+  AfterViewInit,
+  Input,
+  Output,
+  EventEmitter,
+} from '@angular/core';
 import { Router } from '@angular/router';
 
 import * as Feather from 'feather-icons';
@@ -6,38 +13,43 @@ import * as Feather from 'feather-icons';
 @Component({
   selector: 'app-toolbar',
   templateUrl: './toolbar.component.html',
-  styleUrls: ['./toolbar.component.css']
+  styleUrls: ['./toolbar.component.css'],
 })
 export class ToolbarComponent implements OnInit, AfterViewInit {
-
   @Input('whichPage') whichPage: number;
   @Output('toggleSidNav') show: EventEmitter<boolean>;
 
   public opened: boolean;
 
-  constructor(
-    private _router: Router,
-  ) {
+  constructor(private _router: Router) {
     this.whichPage = 0;
     this.show = new EventEmitter<boolean>();
     this.opened = false;
   }
 
-  ngOnInit(): void {
-  }
+  ngOnInit(): void {}
 
   ngAfterViewInit() {
     Feather.replace();
   }
 
   public logOut() {
-    localStorage.clear();
-    this._router.navigate(['/auth/login']);
+    if (this.whichPage != 0) {
+      localStorage.clear();
+      this._router.navigate(['/auth/login']);
+    } else {
+      if (localStorage.getItem('id_rol') == '1') {
+        this._router.navigate(['/soccer-stats/admin']);
+      } else if (localStorage.getItem('id_rol') == '2') {
+        this._router.navigate(['/soccer-stats/employee']);
+      } else {
+        this._router.navigate(['/soccer-stats']);
+      }
+    }
   }
 
   public toggleMenu() {
     this.opened = !this.opened;
     this.show.emit(this.opened);
   }
-
 }
